trait rename to traits
